fix: restore CLI entry point by moving agent summary cache to src
The helpers were exported from cli/translator.mjs so tests could import
them, which required an import.meta.url main-module guard on the entry
point. That guard breaks npx and installed execution, since npm links
package bins and argv[1] then differs from the realpath-resolved
import.meta.url — the CLI silently exits 0 without running. Moving the
helpers to src removes the need for both the export and the guard.
